Move out tools and prompts to separate files for cleaner management
